get stuff working

+++ catkin_ws/src/grabber-controller/scripts/grabber_feedback_subscriber.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 import rospy
-from std_msgs.msg import Float64  # Replace with the correct message type if needed
+from std_msgs.msg import Float32  # Replace with the correct message type if needed
 
 def feedback_callback(msg):
     # Log or process the received feedback data
@@ -12,7 +12,7 @@ def motor_feedback_listener():
     # Initialize the ROS node
     rospy.init_node('motor_feedback_subscriber', anonymous=True)
     # Create a subscriber object
-    rospy.Subscriber('motor_feedback', Float64, feedback_callback)
+    rospy.Subscriber('motor_feedback', Float32, feedback_callback)
     # Keep the node running and listening for messages
     rospy.spin()
